# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VerdictAI: Frictionless Dispute & Chargeback Resolution",
    "description": "\n    # \u2696\ufe0f VerdictAI REST API Gateway\n    **Frictionless Dispute & Chargeback Resolution Platform**\n    \n    *Engineered for Card Members (Mobile), Merchants (Web), and Dispute-Ops Analysts/Admins (Web).*\n    \n    ### API Capabilities:\n    * **Dispute Lifecycle Operations:** Case file creation, query filtering, state machine transitions.\n    * **Polymorphic Evidence Storage:** Ingestion of binary documents (PDF/JPG/PNG) & structured JSON with SHA-256 integrity hashes.\n    * **Dispute-Ops Admin Queue:** Filterable review queue & audited manual overrides.\n    * **Cryptographic Audit Engine:** Tamper-evident hash-chained logs.\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/reasoning/generate": {
      "post": {
        "tags": [
          "Reasoning & XAI"
        ],
        "summary": "Generate explainable reasoning from scoring factors",
        "description": "Direct inference endpoint generating transparent plain-language rationales\n(cardholder, merchant, and objective summary) for arbitrary dispute scoring payloads.\nAll outputs are deterministically audited by HallucinationGuardrail before dispatch.",
        "operationId": "generate_reasoning_api_v1_reasoning_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasoningGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasoningResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reasoning/case/{dispute_id}/generate": {
      "post": {
        "tags": [
          "Reasoning & XAI"
        ],
        "summary": "Generate and persist reasoning for an existing dispute case",
        "description": "Generates, audits, and persists explainable reasoning for an existing dispute case.\nCoordinates between CaseService, FairWeighingScoringService, and PostgreSQL audit chain.",
        "operationId": "generate_case_reasoning_api_v1_reasoning_case__dispute_id__generate_post",
        "parameters": [
          {
            "name": "dispute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dispute Id"
            }
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Actor identity triggering the explanation generation",
              "default": "SYSTEM:reasoning_api",
              "title": "Actor"
            },
            "description": "Actor identity triggering the explanation generation"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasoningResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reasoning/case/{dispute_id}": {
      "get": {
        "tags": [
          "Reasoning & XAI"
        ],
        "summary": "Retrieve stored reasoning for a dispute case",
        "description": "Retrieves previously evaluated explainable reasoning, factor breakdown, and\nguardrail verification status for an existing dispute case.",
        "operationId": "get_case_reasoning_api_v1_reasoning_case__dispute_id__get",
        "parameters": [
          {
            "name": "dispute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dispute Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasoningResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reasoning/audit-guardrails": {
      "post": {
        "tags": [
          "Reasoning & XAI"
        ],
        "summary": "Audit plain-language text against policy guardrails (PII, settlement promises, bias)",
        "description": "Pre-submission audit endpoint enabling web (e.g. Admin override textarea) and mobile clients\nto audit narrative justification text against strict policy rules prior to persistent storage.\nEnforces PII redaction, prohibits unauthorized settlement promises, and detects adversarial bias.",
        "operationId": "audit_text_guardrails_api_v1_reasoning_audit_guardrails_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardrailAuditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardrailAuditResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GuardrailAuditRequest": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text",
            "description": "Plain-language text to audit against policy guardrails",
            "example": "Merchant failed to deliver merchandise per order receipt."
          },
          "expected_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Amount",
            "description": "Expected disputed amount for monetary sanity checking"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "Currency symbol/code",
            "default": "INR"
          },
          "contributing_factors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contributing Factors",
            "description": "Associated factor descriptions"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "GuardrailAuditRequest"
      },
      "GuardrailAuditResponse": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed",
            "description": "True if text strictly complies with all policies"
          },
          "violations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Violations",
            "description": "List of detected policy violations"
          },
          "sanitized_text": {
            "type": "string",
            "title": "Sanitized Text",
            "description": "Scrubbed and sanitized text"
          },
          "checked_rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Checked Rules",
            "description": "List of guardrail verification rules applied"
          },
          "audited_at": {
            "type": "string",
            "format": "date-time",
            "title": "Audited At"
          }
        },
        "type": "object",
        "required": [
          "passed",
          "sanitized_text"
        ],
        "title": "GuardrailAuditResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReasoningGenerateRequest": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id",
            "description": "Unique case reference identifier",
            "example": "CASE-2026-00001"
          },
          "category_id": {
            "type": "string",
            "title": "Category Id",
            "description": "Dispute category identifier (e.g. CAT-01)",
            "default": "CAT-01",
            "example": "CAT-01"
          },
          "category_name": {
            "type": "string",
            "title": "Category Name",
            "description": "Human-readable dispute category name",
            "default": "Item Not Received",
            "example": "Item Not Received"
          },
          "recommended_resolution": {
            "type": "string",
            "title": "Recommended Resolution",
            "description": "CARD_MEMBER_FAVOUR | MERCHANT_FAVOUR | ESCALATE",
            "example": "CARD_MEMBER_FAVOUR"
          },
          "confidence_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confidence Score",
            "description": "Calculated model confidence percentage (0.0 - 100.0)",
            "example": 85.0
          },
          "card_member_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Card Member Score",
            "description": "Normalized card member score (0.0 - 100.0)",
            "default": 0.0,
            "example": 78.5
          },
          "merchant_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Merchant Score",
            "description": "Normalized merchant score (0.0 - 100.0)",
            "default": 0.0,
            "example": 25.0
          },
          "disputed_amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Disputed Amount",
            "description": "Disputed financial transaction amount",
            "example": 1250.0
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "Transaction currency code",
            "default": "INR",
            "example": "INR"
          },
          "factor_breakdown": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Factor Breakdown",
            "description": "Detailed item-level score contributions"
          },
          "raw_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Statement",
            "description": "Cardholder dispute statement"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "recommended_resolution",
          "confidence_score"
        ],
        "title": "ReasoningGenerateRequest"
      },
      "ReasoningResponse": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "Objective, neutral plain-language justification for both parties"
          },
          "cardholder_rationale": {
            "type": "string",
            "title": "Cardholder Rationale",
            "description": "Cardholder-facing plain-language rationale"
          },
          "merchant_rationale": {
            "type": "string",
            "title": "Merchant Rationale",
            "description": "Merchant-facing evidence evaluation rationale"
          },
          "contributing_factors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 3,
            "title": "Contributing Factors",
            "description": "At least 3 factors cited per SRS AC-10"
          },
          "generator_source": {
            "type": "string",
            "title": "Generator Source",
            "description": "GEMINI_XAI | DETERMINISTIC_SAFE_FALLBACK",
            "default": "DETERMINISTIC_SAFE_FALLBACK"
          },
          "guardrails_applied": {
            "type": "boolean",
            "title": "Guardrails Applied",
            "description": "Indicates whether deterministic policy guardrails ran",
            "default": true
          },
          "guardrail_violations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Guardrail Violations",
            "description": "Any caught and mitigated violations"
          },
          "confidence_score_pct": {
            "type": "number",
            "title": "Confidence Score Pct",
            "description": "Confidence score associated with this reasoning"
          },
          "recommended_resolution": {
            "type": "string",
            "title": "Recommended Resolution",
            "description": "Recommended resolution string"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id",
            "description": "Associated case file ID if linked"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At",
            "description": "Timestamp of explanation generation"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "cardholder_rationale",
          "merchant_rationale",
          "contributing_factors",
          "confidence_score_pct",
          "recommended_resolution"
        ],
        "title": "ReasoningResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}